pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vl2d"
version = "0.1.0"
description = "Video to reviewable subtitle-aligned speech dataset pipeline."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [
  { name = "VL2D Contributors" }
]
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.36",
  "alembic>=1.14.0",
  "typer>=0.12.5",
  "pydantic>=2.9.2",
  "python-multipart>=0.0.12",
  "pillow>=11.0.0",
  "httpx>=0.28.0",
  "rich>=13.9.4"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3",
  "pytest-asyncio>=0.24.0"
]
ml = [
  "silero-vad>=5.1.2",
  "torch>=2.5.1"
]
ocr-paddle = [
  "paddleocr>=2.9.1"
]

[project.scripts]
vl2d = "vl2d.cli:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
